fix: edit comment func, style: css homepage

+++ content/view/chapter.php
@@ -1,31 +1,38 @@
 <?php $title = "Jean Forteroche " . $currentChapter->getTitle(); ?>
 
 
-<div id="chapter_page_container"> 
+<div id="chapter_page_container">
     <div id="chapter_container">
         <h1> <?php echo $currentChapter->getTitle(); ?></h1>
         <p><?php echo $currentChapter->getContent(); ?></p>
+        <date><?php echo $currentChapter->getCreateDate(); ?></date>
     </div>
 
-    <div id="comment_container">
+    <div id="comments_container">
         <?php if (isset($_SESSION['id'])) : ?>
-            <form id="form_add_comment_container" action="<?php echo HOST; ?>addComment/id/<?php echo $currentChapter->getId()?>" method="post">
+            <form id="form_add_comment_container" action="<?php echo HOST; ?>addComment/id/<?php echo $currentChapter->getId() ?>" method="post">
                 <textarea id="area_add_comment_container" name='values[content]' placeholder="Votre commentaire"></textarea>
-                <input type="submit" value="Valider"/>
+                <input class='btn' type="submit" value="Valider" />
             </form>
         <?php endif; ?>
         <?php if (isset($comments)) : ?>
             <?php foreach ($comments as $comment) : ?>
                 <div id="comment_container">
-                    <h2><?php echo $comment['pseudo'] ?></h2>
-                    <span><?php echo $comment['create_date'] ?></span>
+                    <h3><?php echo $comment['pseudo'] ?></h3>
                     <?php echo $comment['content'] ?>
-                    <?php if (isset($_SESSION['id'])) : ?>
-                        <span><a href="<?php echo HOST; ?>editComment/id/<?php echo $comment['id'] ?>" class="edit_com_btn btn">Editer</a>
+                    <div class="date_time_comment">
+                        <time>Crée le <?php echo $comment['create_date'] ?></time>
+                        <?php if (isset($comment['edit_date'])) : ?>
+                            <time>Edité le <?php echo $comment['edit_date'] ?></time>
+                        <?php endif; ?>
+                    </div>
+                    <span class='comment_btn_container'>
+                        <?php if (isset($_SESSION['id']) && ($_SESSION['pseudo'] === $comment['pseudo'])) : ?>
+                            <a href="<?php echo HOST; ?>editComment/id/<?php echo $comment['id'] ?>" class="edit_com_btn btn">Editer</a>
                             <a href="<?php echo HOST; ?>deleteComment/id/<?php echo $comment['id'] ?>" class="erase_com_btn btn">Effacer</a>
-                            <a href="<?php echo HOST; ?>reportComment/id/<?php echo $comment['id'] ?>" class="report_com_btn btn">Signaler</a>
-                        </span>
-                    <?php endif; ?>
+                        <?php endif; ?>
+                        <a href="<?php echo HOST; ?>reportComment/id/<?php echo $comment['id'] ?>" class="report_com_btn btn">Signaler</a>
+                    </span>
                 </div>
             <?php endforeach; ?>
         <?php endif; ?>
